@@ -1,9 +1,10 @@
 {
   "name": "@digicatapult/wasp-routing-service",
-  "version": "2.0.4",
+  "version": "2.1.0",
   "description": "",
   "main": "index.js",
   "license": "Apache-2.0",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/digicatapult/wasp-routing-service.git"
@@ -13,35 +14,35 @@
     "build:docker": "./scripts/build-docker.sh",
     "start": "node app/index.js",
     "dev": "nodemon --watch app --watch .env app/index.js | pino-colada",
-    "test": "NODE_ENV=test mocha --config ./test/mocharc.js ./test",
+    "test": "NODE_ENV=test mocha --config ./test/mocharc.json ./test",
     "lint": "eslint .",
     "depcheck": "depcheck",
-    "coverage": "nyc npm run test"
+    "coverage": "c8 npm run test"
   },
   "engines": {
-    "node": ">=14.16.0"
+    "node": ">=18.13.0"
   },
   "dependencies": {
-    "dotenv": "^10.0.0",
-    "envalid": "^7.1.1",
-    "express": "^4.17.1",
-    "kafkajs": "^1.15.0",
-    "node-fetch": "^2.6.1",
-    "pino": "^6.13.0",
-    "pino-http": "^5.6.0"
+    "dotenv": "^16.0.3",
+    "envalid": "^7.3.1",
+    "express": "^4.18.2",
+    "kafkajs": "^2.2.3",
+    "node-fetch": "^3.3.0",
+    "pino": "^8.8.0",
+    "pino-http": "^8.3.1"
   },
   "devDependencies": {
-    "chai": "^4.3.4",
+    "c8": "^7.12.0",
+    "chai": "^4.3.7",
     "delay": "^5.0.0",
-    "depcheck": "^1.4.2",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "mocha": "^9.0.3",
-    "nodemon": "^2.0.12",
-    "nyc": "^15.1.0",
-    "pino-colada": "^2.1.0",
-    "prettier": "^2.3.2",
-    "supertest": "^6.1.4"
+    "depcheck": "^1.4.3",
+    "eslint": "^8.32.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "mocha": "^10.2.0",
+    "nodemon": "^2.0.20",
+    "pino-colada": "^2.2.2",
+    "prettier": "^2.8.3",
+    "supertest": "^6.3.3"
   }
 }
