@@ -1,3 +1,3 @@
-const { startServer } = require('./server')
+import { startServer } from './server.js'
 
 startServer()
