@@ -1,5 +1,5 @@
-const pino = require('pino')
-const env = require('./env')
+import pino from 'pino'
+import env from './env.js'
 
 const logger = pino(
   {
@@ -11,4 +11,4 @@ const logger = pino(
 
 logger.debug('Environment variables: %j', env)
 
-module.exports = logger
+export default logger
