@@ -1,11 +1,12 @@
-const { Kafka, logLevel: kafkaLogLevels } = require('kafkajs')
-const { InvalidIngestIdError } = require('./errors')
+import { Kafka, logLevel as kafkaLogLevels } from 'kafkajs'
+import { InvalidIngestIdError } from './errors.js'
 
-const logger = require('./logger')
-const { KAFKA_BROKERS, KAFKA_PAYLOAD_TOPIC, KAFKA_LOG_LEVEL, KAFKA_PAYLOAD_ROUTING_PREFIX } = require('./env')
-const {
-  things: { getThings },
-} = require('./services')
+import logger from './logger.js'
+import env from './env.js'
+import { things } from './services/index.js'
+
+const { KAFKA_BROKERS, KAFKA_PAYLOAD_TOPIC, KAFKA_LOG_LEVEL, KAFKA_PAYLOAD_ROUTING_PREFIX } = env
+const { getThings } = things
 
 const setupRawPayloadConsumer = async () => {
   const kafkaLogger = logger.child({ module: 'kafkajs-payloads' }, { level: 'error' })
@@ -98,4 +99,4 @@ const setupRawPayloadConsumer = async () => {
   }
 }
 
-module.exports = setupRawPayloadConsumer
+export default setupRawPayloadConsumer
